Update BLEAdvertising.cpp
Apple iBeacon requires a max and min advertising interval of 100ms.  It appears to me we are kind of hard coded at 32ms min and 64ms max.  We could probably do an iBeacon using setManufacturersData() and setFlags() but we need to be able to modify advertising interval to truly accommodate iBeacon spec.

I think this code does it.